requirements.txt
streamlit==1.28.1
openai==1.3.0
PyPDF2==3.0.1
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0